refactor(lotto) : 패키지 위치 조정
- dto 패키지 domain 바깥으로 이동.
