+++ src/main/java/lotto/domain/LottoChecker.java
@@ -1,8 +1,8 @@
 package lotto.domain;
 
-import lotto.domain.dto.LottoNumbers;
-import lotto.domain.dto.LottoRank;
-import lotto.domain.dto.LottoResult;
+import lotto.dto.LottoNumbers;
+import lotto.dto.LottoRank;
+import lotto.dto.LottoResult;
 
 import java.util.Arrays;
 import java.util.List;
