+++ src/test/java/lotto/domain/LottoBuyerTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.dto.LottoNumbers;
+import lotto.dto.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
