+++ src/test/java/lotto/domain/LottoCheckerTest.java
@@ -1,8 +1,8 @@
 package lotto.domain;
 
-import lotto.domain.dto.LottoNumbers;
-import lotto.domain.dto.LottoRank;
-import lotto.domain.dto.LottoResult;
+import lotto.dto.LottoNumbers;
+import lotto.dto.LottoRank;
+import lotto.dto.LottoResult;
 import org.junit.jupiter.api.Test;
 
 import java.util.Collections;
