+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.dto.LottoRank;
+import lotto.dto.LottoRank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
