+++ src/test/java/lotto/dto/LottoRankTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.dto;
+package lotto.dto;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
